[FIX] Added an exact (upper bound) solution to the discrimination experiments in 'tabulate'.
